Move and rename redux types
